LayoutTests/fast/events/before-input-prevent-paste-expected.txt

To manually test this, copy any text and try to paste into the input field. The value of the input should not change.

initial value:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "helloworld"

after the cut:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       <br>

after prevented paste:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       <br>

after allowed paste:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"
|       "helloworld"
|       <br>
